build(fonts): add galada font to project
As I don't have enough skills to create the appropriate images for a possible logo of the
application, as an alternative, the Galada font was added through which with a simple UILabel
component it is possible to reproduce the behavior that would be achieved with the respective image
component. The downside is that the ttf file is considerable bigger than a possible svg file with the logo. Sorry! 😰