fix: remove the main tab bar controller
The initial idea was for the photo of the day to be highlighted. This way, we would have a dedicated
tab for your presentation, which in the initial plan would be the main tab of the tab bar
controller. During the code, I realized that the application's biggest appeal should be on the
infinite scroll list, so I reversed the idea, bringing the tab with the list as the main one and
leaving the Apod tab as secondary. Over time, I realized that it was not fulfilling its original
purpose. Also, the tab bar, especially in the Light-themed presentation, due to the nature of the
photos that mostly have darker tones, was out of place on the main screen, breaking the design. For
these reasons, I decided to remove the tabs, and the app is now just a single view app.